+++ H2TauTau/src/classes_def.xml
@@ -6,5 +6,5 @@
  <class name="std::vector<cmg::METSignificance>" />
  <class name="edm::Wrapper<std::vector<cmg::METSignificance> >" />
  <class name="HTTRecoilCorrector"/>
-  
+ <function name="getDYWeight"/>  
 </lcgdict>
